FUNCIONA PROYECTO

+++ TpFinalProg/Controlador/ProyectoControlador.cs
@@ -9,13 +9,13 @@ using TpFinalProg.Dominio.Entidades;
 namespace TpFinalProg.Controlador {
     internal class ProyectoControlador {
 
-         public static int crear(string nombre, double montoEstimado, int tiempoEstimado, int idPropietario, int legajo) {
+         public static int crear(string nombre, decimal montoEstimado, int tiempoEstimado, int idPropietario, int legajo) {
              Proyecto clsProyecto = new Proyecto(0, nombre, montoEstimado , tiempoEstimado, idPropietario, legajo);
              int idGenerado = clsProyecto.guardar();
              return idGenerado;
          }
 
-        public static void actualizar(int id_proyecto, string nombre, double montoEstimado, int tiempoEstimado, int idPropietario, int legajo) {
+        public static void actualizar(int id_proyecto, string nombre, decimal montoEstimado, int tiempoEstimado, int idPropietario, int legajo) {
              Proyecto p = new Proyecto(id_proyecto, nombre, montoEstimado, tiempoEstimado,  idPropietario, legajo);
              p.guardar();
          }
